+++ yellowstone-grpc-geyser/src/util/testkit.rs
@@ -1,0 +1,20 @@
+use log::{Metadata, Record};
+
+///
+/// A simple logger that logs messages to the console.
+///
+pub struct StdoutLogger;
+
+impl log::Log for StdoutLogger {
+    fn enabled(&self, metadata: &Metadata) -> bool {
+        metadata.level() <= log::max_level()
+    }
+
+    fn log(&self, record: &Record) {
+        if self.enabled(record.metadata()) {
+            println!("{} - {}", record.level(), record.args());
+        }
+    }
+
+    fn flush(&self) {}
+}