+++ yellowstone-grpc-proto/src/lib.rs
@@ -148,6 +148,7 @@ pub mod convert_to {
             loaded_addresses,
             return_data,
             compute_units_consumed,
+            cost_units,
         } = meta;
         let err = create_transaction_error(status);
         let inner_instructions_none = inner_instructions.is_none();
@@ -186,6 +187,7 @@ pub mod convert_to {
             return_data: return_data.as_ref().map(create_return_data),
             return_data_none: return_data.is_none(),
             compute_units_consumed: *compute_units_consumed,
+            cost_units: *cost_units,
         }
     }
 
@@ -497,6 +499,7 @@ pub mod convert_from {
                 })
             },
             compute_units_consumed: meta.compute_units_consumed,
+            cost_units: meta.cost_units,
         })
     }