+++ yellowstone-grpc-proto/src/plugin/filter/filter.rs
@@ -1167,6 +1167,7 @@ mod tests {
             loaded_addresses: LoadedAddresses::default(),
             return_data: None,
             compute_units_consumed: None,
+            cost_units: None,
         });
         let sig = sanitized_transaction.signature();
         let account_keys = sanitized_transaction